+++ xs2a-impl/src/main/java/de/adorsys/aspsp/xs2a/config/ScaAuthorizationConfig.java
@@ -16,17 +16,15 @@
 
 package de.adorsys.aspsp.xs2a.config;
 
+import de.adorsys.aspsp.xs2a.config.factory.AisScaStageAuthorisationFactory;
 import de.adorsys.aspsp.xs2a.service.authorization.ais.*;
 import de.adorsys.aspsp.xs2a.service.authorization.pis.*;
-import de.adorsys.aspsp.xs2a.service.consent.AisConsentDataService;
 import de.adorsys.aspsp.xs2a.service.consent.AisConsentService;
 import de.adorsys.aspsp.xs2a.service.mapper.consent.Xs2aAisConsentMapper;
 import de.adorsys.aspsp.xs2a.service.mapper.consent.Xs2aPisConsentMapper;
-import de.adorsys.aspsp.xs2a.service.mapper.spi_xs2a_mappers.SpiResponseStatusToXs2aMessageErrorCodeMapper;
 import de.adorsys.aspsp.xs2a.service.payment.*;
 import de.adorsys.aspsp.xs2a.service.profile.AspspProfileServiceWrapper;
 import de.adorsys.psd2.aspsp.profile.domain.ScaApproach;
-import de.adorsys.psd2.xs2a.spi.service.AisConsentSpi;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -38,7 +36,6 @@ import static de.adorsys.psd2.aspsp.profile.domain.ScaApproach.*;
 @RequiredArgsConstructor
 public class ScaAuthorizationConfig {
     private final AspspProfileServiceWrapper aspspProfileService;
-    private final AisConsentDataService aisConsentDataService;
 
     @Bean
     public ScaPaymentService scaPaymentService(
@@ -60,16 +57,16 @@ public class ScaAuthorizationConfig {
     }
 
     @Bean
-    public AisAuthorizationService aisAuthorizationService(AisConsentSpi aisConsentSpi, AisConsentService aisConsentService,
-                                                           Xs2aAisConsentMapper aisConsentMapper, SpiResponseStatusToXs2aMessageErrorCodeMapper messageErrorCodeMapper
-    ) {
+    public AisAuthorizationService aisAuthorizationService(AisConsentService aisConsentService,
+                                                           Xs2aAisConsentMapper aisConsentMapper,
+                                                           AisScaStageAuthorisationFactory scaStageAuthorisationFactory) {
         switch (getScaApproach()) {
             case OAUTH:
                 return new OauthAisAuthorizationService();
             case DECOUPLED:
                 return new DecoupledAisAuthorizationService();
             case EMBEDDED:
-                return new EmbeddedAisAuthorizationService(aisConsentService, aisConsentMapper, aisConsentDataService, aisConsentSpi, messageErrorCodeMapper);
+                return new EmbeddedAisAuthorizationService(aisConsentService, aisConsentMapper, scaStageAuthorisationFactory);
             default:
                 return new RedirectAisAuthorizationService();
         }