+++ xs2a-impl/src/main/java/de/adorsys/aspsp/xs2a/service/mapper/spi_xs2a_mappers/Xs2aToSpiPsuDataMapper.java
@@ -14,15 +14,19 @@
  * limitations under the License.
  */
 
-package de.adorsys.psd2.consent.api;
+package de.adorsys.aspsp.xs2a.service.mapper.spi_xs2a_mappers;
 
-public enum CmsScaStatus {
-    RECEIVED,
-    PSUIDENTIFIED,
-    PSUAUTHENTICATED,
-    SCAMETHODSELECTED,
-    STARTED,
-    FINALISED,
-    FAILED,
-    EXEMPTED
+import de.adorsys.aspsp.xs2a.domain.psu.Xs2aPsuData;
+import de.adorsys.psd2.xs2a.spi.domain.psu.SpiPsuData;
+import org.springframework.stereotype.Component;
+
+@Component
+public class Xs2aToSpiPsuDataMapper {
+
+    public SpiPsuData mapToSpiPsuData(Xs2aPsuData xs2aPsuData) {
+        return new SpiPsuData(xs2aPsuData.getPsuId(),
+            xs2aPsuData.getPsuIdType(),
+            xs2aPsuData.getPsuCorporateId(),
+            xs2aPsuData.getPsuCorporateIdType());
+    }
 }