+++ doc/releasenotes.md
@@ -24,3 +24,16 @@
 - Fix the account consent information request to return link information.
 - Add endpoint updateAspspConsentData for PIS consent to Consent Management Service.
 - Remove empty links in the responses.
+
+
+## version 1.9 (planned date: 11.10.2018)
+- Confirmation of funds request. Interface on side of SPI is added.
+- Validation of TPP data is added.
+- Change logic of implicit/explicit method of authorisation
+- AspspConsentData field in the Consent is updated. Support base64 encoding.
+- Validation of Consent (expiration date).
+- Get list of reachable accounts. Embedded approach.
+- Get balances for a given account. Embedded approach.
+- Migration to package and Maven GroupId "de.adorsys.psd2": Consent Management
+- cms-client to work with AspspConsentData endpoints is exended.
+